1318 - move auth folder to client
