+++ src/components/client/about-project/AboutProjectPage.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'next-i18next'
 
 import { Container } from '@mui/material'
 
-import Layout from 'components/layout/Layout'
+import Layout from 'components/client/layout/Layout'
 import Timeline from './sections/TimeLine/Timeline'
 import TechStack from './sections/TechStack/TechStack'
 import WhatIsDone from './sections/WhatIsDone/WhatIsDone'
