+++ src/components/client/about-project/helpers/technologyStackData.tsx
