+++ src/components/client/about-project/helpers/whatIsDoneData.tsx
