+++ src/components/client/about-project/icons/ActivityIcon.tsx
