+++ src/components/client/about-project/icons/GlobeIcon.tsx
