+++ src/components/client/about-project/icons/HandIcon.tsx
