+++ src/components/client/about-project/icons/InvestedHoursIcon.tsx
