+++ src/components/client/about-project/icons/MeetingsIcon.tsx
