+++ src/components/client/about-project/icons/NameSuggestionIcon.tsx
