+++ src/components/client/about-project/icons/RocketIcon.tsx
