+++ src/components/client/about-project/sections/AboutPlatform/AboutPlatform.tsx
