+++ src/components/client/about-project/sections/TimelineItem/TimelineItem.styled.tsx
