+++ src/components/client/about-project/sections/WhatIsDone/WhatIsDone.styled.tsx
