+++ src/components/client/auth/changePassword/ChangePasswordPage.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import { useTranslation } from 'next-i18next'
 import { Container } from '@mui/material'
 
-import Layout from 'components/layout/Layout'
-import ChangePasswordForm from 'components/auth/changePassword/ChangePasswordForm'
+import Layout from 'components/client/layout/Layout'
+import ChangePasswordForm from 'components/client/auth/changePassword/ChangePasswordForm'
 
 export default function ChangePasswordPage() {
   const { t } = useTranslation()
