+++ src/components/client/auth/forgottenPassword/ForgottenPasswordPage.tsx
@@ -2,8 +2,8 @@ import React from 'react'
 import { useTranslation } from 'next-i18next'
 import { Container } from '@mui/material'
 
-import Layout from 'components/layout/Layout'
-import ForgottenPasswordForm from 'components/auth/forgottenPassword/ForgottenPasswordForm'
+import Layout from 'components/client/layout/Layout'
+import ForgottenPasswordForm from 'components/client/auth/forgottenPassword/ForgottenPasswordForm'
 
 export default function ForgottenPasswordPage() {
   const { t } = useTranslation()
